# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Cohort — AI agent fleet governance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /connectors:
    get:
      operationId: listConnectors
      tags: [connectors]
      summary: List platform connectors
      responses:
        "200":
          description: Connectors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Connector"
    post:
      operationId: connectPlatform
      tags: [connectors]
      summary: Connect a platform (plug and play)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectorInput"
      responses:
        "201":
          description: Connector created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connector"
  /connectors/{connectorId}/discover:
    post:
      operationId: discoverAgents
      tags: [connectors]
      summary: Run discovery on a connector
      description: >-
        Scans the connected platform, discovers running agents and proposes a
        5-layer KPI mapping for each so they can be admitted in one click.
      parameters:
        - name: connectorId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Discovery result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoveryResult"
  /connectors/{connectorId}/import:
    post:
      operationId: importDiscoveredAgents
      tags: [connectors]
      summary: Admit discovered agents with their proposed metrics
      parameters:
        - name: connectorId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscoveryImportInput"
      responses:
        "201":
          description: Admitted agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
...
  /connectors/register:
    post:
      operationId: registerConnector
      tags: [connectors]
      summary: Register a real connector with a credential and test it
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterConnectorInput"
      responses:
        "201":
          description: Connector registered (test result included)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterConnectorResult"
  /connectors/{connectorId}/test:
    post:
      operationId: testConnector
      tags: [connectors]
      summary: Test a registered connector's credential against the live platform
      parameters:
        - name: connectorId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorTestResult"
...
components:
  schemas:
...
    Agent:
      type: object
      required:
        - id
        - name
        - slug
        - role
        - platform
        - version
        - status
        - bio
        - currentVerdict
        - verdictConfidence
        - severity
        - healthScore
        - admittedAt
        - lastEvaluatedAt
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        role:
          type: string
        platform:
          type: string
        version:
          type: string
        status:
          type: string
          enum: [observation, active, flagged, retiring, retired]
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: string
        tagline:
          type: string
        monthlyVolume:
          type: integer
        headlineKpis:
          type: array
          items:
            $ref: "#/components/schemas/KpiMetric"
        targetMetrics:
          type: array
          description: >-
            Flattened metrics from the agent's latest evaluation that carry a
            comparable goal. Used by the fleet view to roll up how many metrics
            are currently off-target.
          items:
            $ref: "#/components/schemas/KpiMetric"
        currentVerdict:
          type: string
          enum: [promote, mentor, retire, observation]
        verdictConfidence:
          type: number
        severity:
          type: string
          enum: [critical, high, medium, stable]
        healthScore:
          type: number
        activeAlerts:
          type: integer
        monthlyValue:
          type: number
        monthlyCost:
          type: number
        admittedAt:
          type: string
        lastEvaluatedAt:
          type: string
...
    KpiMetric:
      type: object
      required:
        - label
        - value
        - unit
        - trend
      properties:
        label:
          type: string
        value:
          type: number
        unit:
          type: string
        trend:
          type: number
        direction:
          type: string
          enum: [up, down, flat]
        target:
          type: string
        rationale:
          type: string
...
    Connector:
      type: object
      required:
        - id
        - platform
        - name
        - status
        - agentsDiscovered
        - category
      properties:
        id:
          type: string
        platform:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [connected, available, syncing]
        agentsDiscovered:
          type: integer
        category:
          type: string
        lastSyncAt:
          type: ["string", "null"]
...
    ConnectorInput:
      type: object
      required:
        - platform
      properties:
        platform:
          type: string
        name:
          type: string
    ProposedMetric:
      type: object
      required:
        - layer
        - label
        - sourceSignal
        - value
        - unit
        - confidence
      properties:
        layer:
          type: string
          enum: [efficacy, efficiency, adoption, governance, value]
        label:
          type: string
        sourceSignal:
          type: string
        value:
          type: number
        unit:
          type: string
        confidence:
          type: number
    DiscoveredAgent:
      type: object
      required:
        - externalId
        - name
        - role
        - platform
        - signals
        - proposedMetrics
        - confidence
        - alreadyImported
      properties:
        externalId:
          type: string
        name:
          type: string
        role:
          type: string
        platform:
          type: string
        signals:
          type: array
          items:
            type: string
        proposedMetrics:
          type: array
          items:
            $ref: "#/components/schemas/ProposedMetric"
        proposedVerdict:
          type: string
          enum: [promote, mentor, retire, observation]
        confidence:
          type: number
        alreadyImported:
          type: boolean
    DiscoveryResult:
      type: object
      required:
        - connectorId
        - platform
        - discoveredAt
        - agentsFound
        - agents
        - coverageNote
      properties:
        connectorId:
          type: string
        platform:
          type: string
        discoveredAt:
          type: string
        agentsFound:
          type: integer
        agents:
          type: array
          items:
            $ref: "#/components/schemas/DiscoveredAgent"
        coverageNote:
          type: string
    DiscoveryImportInput:
      type: object
      required:
        - externalIds
      properties:
        externalIds:
          type: array
          items:
            type: string
...
    ConnectorTestResult:
      type: object
      required: [ok, message]
      properties:
        ok:
          type: boolean
        message:
          type: string
    RegisterConnectorInput:
      type: object
      required: [platform, name]
      properties:
        platform:
          type: string
          description: Platform key (e.g. github). Must have a real connector implementation.
        name:
          type: string
          minLength: 1
        token:
          type: ["string", "null"]
          description: Credential (PAT/API key). Omit to use ambient env credentials.
    RegisterConnectorResult:
      type: object
      required: [connector, test]
      properties:
        connector:
          $ref: "#/components/schemas/Connector"
        test:
          $ref: "#/components/schemas/ConnectorTestResult"
...